\name{coord2comp}
\alias{coord2comp}
\title{
Generates secondary structure components from a coordinate file
}
\description{
Split RNA secondary structures into closed components that can be used for 3D assembly into larger molecules. 
}
\usage{
coord2comp(coord)
}
\arguments{
  \item{coord}{
coordinates generated from ct2coord 
}
}
\value{
Data frame containing substructures
}
\author{
JP Bida
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ RSIM }
\keyword{ PDB }% __ONLY ONE__ keyword per line
