\name{ct2ss}
\alias{ct2ss}
\title{
Generates secondary structure components from a coordinate file
}
\description{
Split RNA secondary structures into closed components that can be used for 3D assembly into larger molecules. 
}
\usage{
ct2ss(dat, cleanup)
}
\arguments{
  \item{dat}{
Dataframe from ctfile generated by loadCT
}
  \item{cleanup}{
TRUE/FALSE indicating if CT file should be cleaned up removing pseudoknots and multiple pairings
preserving the most base-pairs in a structure compatible with bracket notation. 
}
}
\value{
Secondary structure in bracket notation
}
\author{
JP Bida
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ RNA }
\keyword{ RSIM }
\keyword{ PDB }% __ONLY ONE__ keyword per line
