\name{genDB}
\alias{genDB}
\title{
Internal Function for creating component database
}
\description{
Adds BREAK to bracket notation in secondary structure indicating 5' and 3' discontinuities at a base-pair
}
\usage{
genDB(map)
}
\arguments{
  \item{map}{
Dataframe generated by a custom script  
}
}
\value{
Returns inputed map with breaks indicated
}
\author{
JP Bida
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ RNA }
\keyword{ RSIM }
\keyword{ PDB }% __ONLY ONE__ keyword per line
