\name{translate}
\alias{translate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
internal function for translating points
}
\description{
internal function to translate points
}
\usage{
translate(x1, y1, x2, y2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x1}{
x coordinates being translated
}
  \item{y1}{
y coordinates being translated
}
  \item{x2}{
dx for translation
}
  \item{y2}{
dy for translation
}
}
\value{
set of points
}
\author{
JP Bida
}
\examples{
## Internal Function ##
}
\keyword{ translate }
